Don't use worker object in Synchronous
